+++ NEWS.md
@@ -1,10 +1,14 @@
 ## fastNaiveBayes 1.0.1
 
 ### New Features
+-
 ### Bug Fixes
+- Fixed bug in Gaussian predict function.
+
 ### Other Changes
-* Changed Readme
-* Changed description
+- Changed Readme
+- Changed description
+- Added unit tests and Travis-ci
 
 ## fastNaiveBayes 1.0.0